backend/requirements.txt
# ============================================================================
# Acadia Log IQ v2.0 — Backend (FastAPI + Bedrock + Hybrid Search)
# ============================================================================

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# AWS Bedrock (Mistral 7B + Titan Embed V2)
boto3>=1.34.0
botocore>=1.34.0

# Vector Store
chromadb>=0.5.5

# Validation
pydantic>=2.7.0,<3
pydantic-settings>=2.3.0,<3

# Rate Limiting
slowapi>=0.1.9

# PDF Extraction
PyMuPDF>=1.24.0

# DOCX Extraction
python-docx>=1.1.0

# Clerk JWT Verification
PyJWT[crypto]>=2.8.0
cryptography>=42.0.0

fastapi
uvicorn[standard]
python-multipart

boto3
botocore

pydantic
pydantic-settings

slowapi

PyMuPDF
python-docx

PyJWT[crypto]
cryptography

sqlalchemy
psycopg[binary]

pgvector
alembic
rank-bm25
slowapi
